chore(pkg/tbtree): file re-organization
