+++ pkg/tbtree/snapshot.go
@@ -1,0 +1,129 @@
+/*
+Copyright 2019-2020 vChain, Inc.
+
+Licensed under the Apache License, Version 2.0 (the "License");
+you may not use this file except in compliance with the License.
+You may obtain a copy of the License at
+
+	http://www.apache.org/licenses/LICENSE-2.0
+
+Unless required by applicable law or agreed to in writing, software
+distributed under the License is distributed on an "AS IS" BASIS,
+WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
+See the License for the specific language governing permissions and
+limitations under the License.
+*/
+package tbtree
+
+import (
+	"errors"
+	"sync"
+)
+
+var ErrReadersNotClosed = errors.New("readers not closed")
+
+type Snapshot struct {
+	t           *TBtree
+	id          uint64
+	root        node
+	readers     map[int]*Reader
+	maxReaderID int
+	closed      bool
+	rwmutex     sync.RWMutex
+}
+
+func (s *Snapshot) Get(key []byte) (value []byte, ts uint64, err error) {
+	s.rwmutex.RLock()
+	defer s.rwmutex.RUnlock()
+
+	if s.closed {
+		return nil, 0, ErrAlreadyClosed
+	}
+
+	return s.root.get(key)
+}
+
+func (s *Snapshot) Ts() (uint64, error) {
+	s.rwmutex.RLock()
+	defer s.rwmutex.RUnlock()
+
+	if s.closed {
+		return 0, ErrAlreadyClosed
+	}
+
+	return s.root.ts(), nil
+}
+
+func (s *Snapshot) Reader(spec *ReaderSpec) (*Reader, error) {
+	s.rwmutex.RLock()
+	defer s.rwmutex.RUnlock()
+
+	if s.closed {
+		return nil, ErrAlreadyClosed
+	}
+
+	if spec == nil {
+		return nil, ErrIllegalArgument
+	}
+
+	path, startingLeaf, startingOffset, err := s.root.findLeafNode(spec.initialKey, nil, nil, spec.ascOrder)
+	if err == ErrKeyNotFound {
+		return nil, ErrNoMoreEntries
+	}
+	if err != nil {
+		return nil, err
+	}
+
+	reader := &Reader{
+		snapshot:   s,
+		id:         s.maxReaderID,
+		initialKey: spec.initialKey,
+		isPrefix:   spec.isPrefix,
+		ascOrder:   spec.ascOrder,
+		path:       path,
+		leafNode:   startingLeaf,
+		offset:     startingOffset,
+		closed:     false,
+	}
+
+	s.readers[reader.id] = reader
+
+	s.maxReaderID++
+
+	return reader, nil
+}
+
+func (s *Snapshot) Close() error {
+	s.rwmutex.Lock()
+	defer s.rwmutex.Unlock()
+
+	if s.closed {
+		return ErrAlreadyClosed
+	}
+
+	if len(s.readers) > 0 {
+		return ErrReadersNotClosed
+	}
+
+	err := s.t.snapshotClosed(s)
+	if err != nil {
+		return err
+	}
+
+	s.closed = true
+
+	return nil
+}
+
+func (s *Snapshot) closedReader(r *Reader) error {
+	s.rwmutex.Lock()
+	defer s.rwmutex.Unlock()
+
+	if s.closed {
+		return ErrAlreadyClosed
+	}
+
+	delete(s.readers, r.id)
+
+	return nil
+}
